+++ common/config/web.php
@@ -34,25 +34,6 @@ $config = [
     ],
 ];
 
-if (YII_ENV_DEV) {
-    // configuration adjustments for 'dev' environment
-    $config['bootstrap'][] = 'debug';
-    $config['modules']['debug'] = [
-        'class'=>'yii\debug\Module',
-        'allowedIPs' => ['*'],
-        'panels'=>[
-            'xhprof'=>[
-                'class'=>'\trntv\debug\xhprof\panels\XhprofPanel'
-            ]
-        ]
-    ];
-
-    $config['bootstrap'][] = 'gii';
-    $config['modules']['gii'] = [
-        'class'=>'yii\gii\Module'
-    ];
-}
-
 $config = yii\helpers\ArrayHelper::merge($base, $config);
 
 return $config;