+++ frontend/config/web.php
@@ -17,4 +17,28 @@ $config = \yii\helpers\ArrayHelper::merge(
     ]
 );
 
+if (YII_ENV_DEV) {
+    // configuration adjustments for 'dev' environment
+    $config['bootstrap'][] = 'debug';
+    $config['modules']['debug'] = [
+        'class'=>'yii\debug\Module',
+        'allowedIPs' => ['*'],
+        'panels'=>[
+            'xhprof'=>[
+                'class'=>'\trntv\debug\xhprof\panels\XhprofPanel'
+            ]
+        ]
+    ];
+
+    $config['bootstrap'][] = 'gii';
+    $config['modules']['gii'] = [
+        'class'=>'yii\gii\Module',
+        'generators'=>[
+            'crud'=>[
+                'class'=>'yii\gii\generators\crud\Generator',
+                'messageCategory'=>'frontend'
+            ]
+        ]
+    ];
+}
 return $config;