+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Yii Framework 2 Change Log
 
 1.4.0 under development
 -----------------------
+- Enh: Gii Module has separate configs for backend and frontend
+- Enh: Added gii templates for backend
 - Enh: Time information on "system information" screen
 - Fix #12: Locale bug
 - Enh: I18N validation rules